+++ Model/Rate.cs
@@ -4,5 +4,5 @@ public class Rate
 {
     public string currency { get; set; }
     public string code { get; set; }
-    public float mid { get; set; }
+    public decimal mid { get; set; }
 }
